slice to integer

+++ random_data.py
@@ -1,6 +1,5 @@
 import numpy as np
 from scipy import sparse 
-import cPickle as pickle
 
 '''
 Generate a random copy of data
@@ -36,8 +35,8 @@ def rand_data():
     
     feature = np.random.rand(n_rows, n_feat)
     
-    trainset = dict(scores=sparse.csr_matrix(score_mat[0:(n_rows / 2)]), atts=feature[0:(n_rows / 2)]) 
-    testset = dict(scores=sparse.csr_matrix(score_mat[(n_rows / 2):]), atts=feature[(n_rows / 2):]) 
+    trainset = dict(scores=sparse.csr_matrix(score_mat[0:int(n_rows / 2)]), atts=feature[0:int(n_rows / 2)]) 
+    testset = dict(scores=sparse.csr_matrix(score_mat[int(n_rows / 2):]), atts=feature[int(n_rows / 2):]) 
 
     return dict(trainset=trainset, testset=testset)
